+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Sync: Fixed error in my account when no email module is available
 - Core/ tickets: fixed error when tickets option "Show confirm dialog when closing tickets" was enabled
 - Core: Fixed user export / import
+- Files: Fixed reload issue after creating new folder for an item
 
 27-01-2020 6.4.107
 - Email: fixed missing email from template option